+++ contracts/interfaces/controller/IControllerState.sol
@@ -27,9 +27,10 @@ interface IControllerState {
     /// @param token 要兑换的代币
     function harvestPath(address token) external view returns (bytes memory);
 
-    /// @notice 涉及到swap时最大滑点 0-1e4，这里是sqrt值
-    function maxSqrtSlippage() external view returns (uint16);
+    /// @notice 获取swap时最大滑点，取值范围为 0-1e4, 计算公式为：MaxSwapSlippage = (1 - (sqrtSlippage/1e4)^2) * 100%
+    ///         如设置最大滑点 0.5%, 则 sqrtSlippage 应设置为9974，此时 MaxSwapSlippage = (1-(9974/1e4)^2)*100% = 0.5% 
+    function maxSqrtSlippage() external view returns (uint32);
 
-    /// @notice 涉及到swap时最大价格影响 0-1e4
-    function maxPriceImpact() external view returns (uint16);
+    /// @notice 获取swap时最大价格影响，取值范围为 0-1e4
+    function maxPriceImpact() external view returns (uint32);
 }
